Refactoring of Camera upload as Sync Provider
This patch reworks most of the camera upload. It contains two important
changes:

(1) The old code implemented camera upload as a service. While that worked,
it proved somewhat problematic. Services can be killed by Android, and the
service had to be started by an Activity.

Android provides an API for implementing data syncing:
https://developer.android.com/training/sync-adapters/index.html

This API provides a number of advantages compared to a self-implemented
service. Among them are:

* Less code
* Automatic error recovery (e.g. retry on network errors)
* Periodic synchronisation (e.g. every 24h)
* Better visibility to the user (via Android Settings->Accounts)

(2) The old camera media sync algortihm tried to upload media via two ways:

a) when a new photo was added to the MediaStore, that photo was uploaded. But
*only* if the data plan allowed it.
b) when the user interacts with Seadroid, the service uploaded all photos
not yet uploaded. It did that by scanning certain hard-coded directories
on the externalStorage  and looking for new media files.

Important changes added by this patch:

- This new design builds soly on the MediaStore. The MediaStore contains
all images/videos on the phone. Therefore it should not be necessary to
scan (or even be aware of) where the photos are actually stored.

- The MediaStore has a column where it records the "DATA_ADDED" for every
media file. We can use that to sort and filter for new newly added media
files. This eases up the syncing mechanism considerably.

- Allowing to sync to sub directories in a repository.

- There is still a service necessary to monitor MediaStore activity. However
even without that service, Android will perform a sync about every 24h.
Also that service is now autostarted on System boot and on Seadroid
apk upgrades.

- Finally merging with existing files on the server was also improved by
skipping files that are already present on the server (it compares the file
size and will still upload if the sizes differ. That's probably the best
one can do with the current API2).

Overall, this patch should adresses the following issues:

issue #277 (Camera upload function dumps all subdir files into single dir)
issue #301 (Inconsistency in camera upload sync)
issue #311 (second suggestion: target repo sub directories)
issue #336 (file duplication when doing a full resync)
issue #375 (Picture upload not acting as service?)
issue #423 (Pictures never upload) [maybe. not sure what's going on there]

Useful links:
https://developer.android.com/training/sync-adapters/creating-sync-adapter.html
http://blog.udinic.com/2013/07/24/write-your-own-android-sync-adapter/